requirements.txt
# Core scientific computing
numpy>=1.26,<2.0
pandas>=2.1
scikit-learn>=1.4
scipy>=1.11

# PyTorch (install separately for CUDA support)
# See: https://pytorch.org/get-started/locally/
torch>=2.1
torchvision>=0.16

# Data I/O
h5py>=3.10
pyarrow>=14.0
pyyaml>=6.0

# Visualization
matplotlib>=3.8
seaborn>=0.13

# Utilities
tqdm>=4.66
joblib>=1.3

# Spatial biology
scanpy>=1.9
anndata>=0.10

# Machine learning
xgboost>=2.0

# Vision models
timm>=0.9

# HuggingFace ecosystem
huggingface-hub>=0.20
transformers>=4.36
datasets>=2.16

# Statistics
statsmodels>=0.14

# HTTP requests (UniProt/GO queries in gene_annotations.py)
requests>=2.31

# Image processing (figure generation)
Pillow>=10.0
